+++ crates/vm/levm/src/utils.rs
@@ -4,7 +4,7 @@ use crate::{
     call_frame::CallFrameBackup,
     constants::*,
     db::gen_db::GeneralizedDatabase,
-    errors::{ExceptionalHalt, InternalError, TxValidationError, VMError},
+    errors::{DatabaseError, ExceptionalHalt, InternalError, TxValidationError, VMError},
     gas_cost::{
         self, ACCESS_LIST_ADDRESS_COST, ACCESS_LIST_STORAGE_KEY_COST, BLOB_GAS_PER_BLOB,
         COLD_ADDRESS_ACCESS_COST, CREATE_BASE_COST, STANDARD_TOKEN_COST,
@@ -19,7 +19,7 @@ use bytes::Bytes;
 use ethrex_common::{
     Address, H256, U256,
     evm::calculate_create_address,
-    types::{Account, Fork, Transaction, tx_fields::*},
+    types::{Account, Code, Fork, Transaction, account_diff::AccountStateDiff, tx_fields::*},
     utils::{keccak, u256_to_big_endian},
 };
 use ethrex_common::{types::TxKind, utils::u256_from_big_endian_const};
@@ -30,7 +30,7 @@ use secp256k1::{
     ecdsa::{RecoverableSignature, RecoveryId},
 };
 use sha3::{Digest, Keccak256};
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
 pub type Storage = HashMap<U256, H256>;
 
 // ================== Address related functions ======================
@@ -164,6 +164,99 @@ pub fn restore_cache_state(
     Ok(())
 }
 
+/// Returns the state diffs introduced by the transaction by comparing the call frame backup
+/// (which holds the state before executing the transaction) with the current state of the cache
+/// (which contains all the writes performed by the transaction).
+pub fn get_account_diffs_in_tx(
+    db: &GeneralizedDatabase,
+    transaction_backup: CallFrameBackup,
+) -> Result<HashMap<Address, AccountStateDiff>, VMError> {
+    let mut modified_accounts = HashMap::new();
+
+    // First we add the account info
+    for (address, original_account) in transaction_backup.original_accounts_info.iter() {
+        let new_account = db
+            .current_accounts_state
+            .get(address)
+            .ok_or(DatabaseError::Custom("DB Cache".to_owned()))?;
+
+        let nonce_diff: u16 = new_account
+            .info
+            .nonce
+            .checked_sub(original_account.info.nonce)
+            .ok_or(InternalError::TypeConversion)?
+            .try_into()
+            .map_err(|_| InternalError::TypeConversion)?;
+
+        let new_balance = if new_account.info.balance != original_account.info.balance {
+            Some(new_account.info.balance)
+        } else {
+            None
+        };
+
+        let bytecode = if new_account.info.code_hash != original_account.info.code_hash {
+            // After execution the code should be in db.codes
+            let code = db
+                .codes
+                .get(&new_account.info.code_hash)
+                .ok_or_else(|| DatabaseError::Custom("Code DB Cache".to_owned()))?;
+            Some(code.clone())
+        } else {
+            None
+        };
+
+        let account_state_diff = AccountStateDiff {
+            new_balance,
+            nonce_diff,
+            storage: BTreeMap::new(), // We add the storage later
+            bytecode: bytecode.map(|c| c.bytecode),
+            bytecode_hash: None,
+        };
+
+        modified_accounts.insert(*address, account_state_diff);
+    }
+
+    // Then if there is any storage change, we add it to the account state diff
+    for (address, original_storage_slots) in
+        transaction_backup.original_account_storage_slots.iter()
+    {
+        let account_info = db
+            .current_accounts_state
+            .get(address)
+            .ok_or(DatabaseError::Custom("DB Cache".to_owned()))?;
+
+        let mut added_storage = BTreeMap::new();
+        for key in original_storage_slots.keys() {
+            added_storage.insert(
+                *key,
+                *account_info
+                    .storage
+                    .get(key)
+                    .ok_or(DatabaseError::Custom("Account info Storage".to_owned()))?,
+            );
+        }
+        if let Some(account_state_diff) = modified_accounts.get_mut(address) {
+            account_state_diff.storage = added_storage;
+        } else {
+            // If the account is not in the modified accounts, we create a new one
+            let account_state_diff = AccountStateDiff {
+                new_balance: None,
+                nonce_diff: 0,
+                storage: added_storage,
+                bytecode: None,
+                bytecode_hash: None,
+            };
+
+            // If account state diff is NOT empty
+            if account_state_diff != AccountStateDiff::default() {
+                modified_accounts.insert(*address, account_state_diff);
+            }
+        }
+    }
+
+    Ok(modified_accounts)
+}
+
 // ================= Blob hash related functions =====================
 pub fn get_base_fee_per_blob_gas(
     block_excess_blob_gas: Option<U256>,
@@ -199,8 +292,8 @@ pub fn get_max_blob_gas_price(
 
     Ok(max_blob_gas_cost)
 }
-/// Gets the actual blob gas cost.
-pub fn get_blob_gas_price(
+/// Calculate the actual blob gas cost.
+pub fn calculate_blob_gas_cost(
     tx_blob_hashes: &[H256],
     block_excess_blob_gas: Option<U256>,
     evm_config: &EVMConfig,
@@ -210,14 +303,14 @@ pub fn get_blob_gas_price(
         .try_into()
         .map_err(|_| InternalError::TypeConversion)?;
 
-    let blob_gas_price: u64 = blobhash_amount
+    let blob_gas_used: u64 = blobhash_amount
         .checked_mul(BLOB_GAS_PER_BLOB)
         .unwrap_or_default();
 
     let base_fee_per_blob_gas = get_base_fee_per_blob_gas(block_excess_blob_gas, evm_config)?;
 
-    let blob_gas_price: U256 = blob_gas_price.into();
-    let blob_fee: U256 = blob_gas_price
+    let blob_gas_used: U256 = blob_gas_used.into();
+    let blob_fee: U256 = blob_gas_used
         .checked_mul(base_fee_per_blob_gas)
         .ok_or(InternalError::Overflow)?;
 
@@ -328,7 +421,7 @@ pub fn eip7702_get_code(
     db: &mut GeneralizedDatabase,
     accrued_substate: &mut Substate,
     address: Address,
-) -> Result<(bool, u64, Address, Bytes), VMError> {
+) -> Result<(bool, u64, Address, Code), VMError> {
     // Address is the delgated address
     let bytecode = db.get_account_code(address)?;
 
@@ -336,13 +429,13 @@ pub fn eip7702_get_code(
     // return false meaning that is not a delegation
     // return the same address given
     // return the bytecode of the given address
-    if !code_has_delegation(bytecode)? {
+    if !code_has_delegation(&bytecode.bytecode)? {
         return Ok((false, 0, address, bytecode.clone()));
     }
 
     // Here the address has a delegation code
     // The delegation code has the authorized address
-    let auth_address = get_authorized_address_from_code(bytecode)?;
+    let auth_address = get_authorized_address_from_code(&bytecode.bytecode)?;
 
     let access_cost = if accrued_substate.add_accessed_address(auth_address) {
         WARM_ADDRESS_ACCESS_COST
@@ -389,8 +482,8 @@ impl<'a> VM<'a> {
             self.substate.add_accessed_address(authority_address);
 
             // 5. Verify the code of authority is either empty or already delegated.
-            let empty_or_delegated =
-                authority_code.is_empty() || code_has_delegation(authority_code)?;
+            let empty_or_delegated = authority_code.bytecode.is_empty()
+                || code_has_delegation(&authority_code.bytecode)?;
             if !empty_or_delegated {
                 continue;
             }
@@ -424,7 +517,7 @@ impl<'a> VM<'a> {
             } else {
                 Bytes::new()
             };
-            self.update_account_bytecode(authority_address, code)?;
+            self.update_account_bytecode(authority_address, Code::from_bytecode(code))?;
 
             // 9. Increase the nonce of authority by one.
             self.increment_account_nonce(authority_address)
@@ -526,7 +619,7 @@ impl<'a> VM<'a> {
     pub fn get_min_gas_used(&self) -> Result<u64, VMError> {
         // If the transaction is a CREATE transaction, the calldata is emptied and the bytecode is assigned.
         let calldata = if self.is_create()? {
-            &self.current_call_frame.bytecode
+            &self.current_call_frame.bytecode.bytecode
         } else {
             &self.current_call_frame.calldata
         };
@@ -580,7 +673,7 @@ impl<'a> VM<'a> {
 
 /// Converts Account to LevmAccount
 /// The problem with this is that we don't have the storage root.
-pub fn account_to_levm_account(account: Account) -> (LevmAccount, Bytes) {
+pub fn account_to_levm_account(account: Account) -> (LevmAccount, Code) {
     (
         LevmAccount {
             info: account.info,